Refactor the software
This PR takes the event handling functionality out of the `canvas` package into a new `controller` package.

Also, it takes the notifier out of the `canvas` package to its own package.

All HTTP handler are now in the `webapp` package.

In the webapp, added a new button to pick the color from the position of the cursor.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>